+++ lib/extract_overall_cost_and_emissions.rb
@@ -1,4 +1,4 @@
-require_relative '../../lib/gdx'
+require_relative 'gdx'
 
 class ExtractOverallCostAndEmissions